+++ ubo_app/store/ethernet.py
@@ -1,0 +1,12 @@
+# ruff: noqa: D100, D101, D102, D103, D104, D107
+from __future__ import annotations
+
+from enum import StrEnum
+
+
+class GlobalEthernetState(StrEnum):
+    CONNECTED = 'Connected'
+    DISCONNECTED = 'Disconnected'
+    PENDING = 'Pending'
+    NEEDS_ATTENTION = 'Needs Attention'
+    UNKNOWN = 'Unknown'